requirements.txt
pytest==7.0.1
# Uncomment if not building in tensorflow docker container
# tensorflow==2.7.1
tqdm==4.62.3
matplotlib==3.5.1
typer==0.4.1
uncertainty-wizard==0.2.0
pandas==1.4.1
# Used within pandas
jinja2==3.0.2
scikit-learn==1.0.2
psutil==5.9.0

foolbox==3.3.2


tensorflow-datasets==4.6.0
protobuf==3.20.0

datasets==2.1.0